pyproject.toml
[project]
name = "smtp-tlsa-verify"
version = "0.1.0"
description = "A small library, command-line utility and JSON/HTTP/FastAPI service to check DANE/TLSA-RRs against SMTP servers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
]

[project.optional-dependencies]
fastapi = [
    "fastapi[standard]>=0.116.1",
]

[dependency-groups]
dev = [
    "pipx>=1.7.1",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
    "tox>=4.28.3",
]

[project.scripts]
danesmtp = "smtp_tlsa_verify.cli:main"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
